move: partner/data-source -> common/data-sources
